Update Password documentation examples
