+++ docs/src/pages/components/cdk-components/password-input/examples/basicPasswordInput.js
@@ -1,0 +1,25 @@
+import { DxcPasswordInput } from "@dxc-technology/halstack-react";
+
+const code = `() => {
+  return (
+    <div style={{ display: "flex" }}>
+      <DxcPasswordInput
+        label="Invalid"
+        clearable
+        margin="medium"
+      />
+      <DxcPasswordInput
+        label="Invalid"
+        error="Error message."
+        clearable
+        margin="medium"
+      />
+    </div>
+  );
+}`;
+
+const scope = {
+  DxcPasswordInput,
+};
+
+export default { code, scope };
